- New LoD methods integrated:   - Tiny LoD: fast + low-memory voxel-based LoD method. Faster than quick-lod and should support 30M+ splats LoDing-on-load   - Bhatt(acharyya) LoD: slower higher quality LoD method using splat shape+color similarity. Used in offline build-lod, also experimental support in the browser using lod="quality"   - (re-computing with .computeLodSplats not yet updated) - Improved efficiency of decoding into PackedSplats/ExtSplats - Commented out experimental LoD bone tree - Integrated splatEncoding ranges into compact Csplat and PackedSplat encodings - New RAD file format encoding+decoding with seekable chunks (streaming/paged not yet implemented) - More unification of Csplat and Gsplat functionality under Tsplat - NewSparkRenderer:   - Added .pagedExtSplats (extSplat encoding for paged/streaming LoDs),   - .accumExtSplats (extSplat encoding for intermediary accumulated splats)   - Added .numLodFetchers (# simultaneous LoD chunk fetchers)   - Added predictive LoD by extrapolating camera motion to where it will be when LoD traversal is complete   - Revised standard LoD splat budget: Oculus (500K) VisionPro (750K), Android (1M), iOS (1.5M), desktop (2.5M). Fixed bug that would incorrectly detect VisionPro in the past. - SplatSkinning: Implemented linear blend skinning for Gsplats in additional dual-quaternion (experimental)
Additional details in GPT-produced summary of diff:
- Added a full RAD format implementation (rad.rs) with encoder/decoder, chunked stream support, and a new SplatFileType::RAD path in decoder.rs.
- Introduced a new Bhattacharyya‑based LOD algorithm (bhatt_lod.rs) and wired it into Rust, WASM, and worker code paths alongside “tiny” LOD.
- Expanded splat encoding plumbing: SplatEncoding/SetSplatEncoding are now serde‑serializable, passed into decoders, and optionally propagated via SplatGetter.
- Added CsplatArray to WASM API with encoding‑aware packed/ext conversion and LOD generation (spark-internal-rs/src/lib.rs).
- Reworked packed/ext splat data buffers in Rust (packed_splats.rs, ext_splats.rs) to cache subranges and reduce JS<->WASM array traffic; LOD child data now uses TSplat APIs.
- Paged splats can now use “ext” encoding (SplatPager.ts): new ext texture, SH3 split textures, readSplatExt/readSplatExtDir, and ext‑aware upload paths.
- Packed splats now default to DEFAULT_SPLAT_ENCODING, SH evaluation uses mid/scale, and LOD creation supports a quality path (tiny vs bhatt); bone‑splats logic is commented out (PackedSplats.ts, SplatLoader.ts).
- Worker pipeline updated: decode to csplats with encoding, tiny_lod_packedsplats/bhatt_lod_packedsplats, and new qualityLodPackedSplats entrypoints (newWorker.ts).
- Renderer LOD logic updated (NewSparkRenderer.ts): accumExtSplats rename, new pagedExtSplats and numLodFetchers, predicted view traversal, and platform‑specific default splat counts.
- Shader/utility tweaks: covariance encoding switches to linear correlation, added quaternionToMatrix helper, and isVisionPro() now requires isIos() (splatDefines.glsl, utils.ts).

+++ rust/spark-lib/src/spz.rs
@@ -628,14 +628,14 @@ fn read_u16_le(two: &[u8]) -> u16 {
 
 pub struct SpzEncoder<T: SplatGetter> {
     getter: T,
-    max_sh_out: Option<u8>,
+    max_sh_out: Option<usize>,
     fractional_bits: u8,
 }
 
 impl<T: SplatGetter> SpzEncoder<T> {
     pub fn new(getter: T) -> Self { Self { getter, max_sh_out: None, fractional_bits: 12 } }
 
-    pub fn with_max_sh(mut self, max_sh: u8) -> Self {
+    pub fn with_max_sh(mut self, max_sh: usize) -> Self {
         self.max_sh_out = Some(max_sh.min(3));
         self
     }
@@ -647,7 +647,7 @@ impl<T: SplatGetter> SpzEncoder<T> {
 
     pub fn encode(mut self) -> anyhow::Result<Vec<u8>> {
         let num_splats = self.getter.num_splats();
-        let sh_src = self.getter.max_sh_degree() as u8;
+        let sh_src = self.getter.max_sh_degree();
         let sh_degree = self.max_sh_out.map(|m| m.min(sh_src)).unwrap_or(sh_src);
         let fractional_bits = self.fractional_bits;
         let flag_antialias = self.getter.flag_antialias();
@@ -659,7 +659,7 @@ impl<T: SplatGetter> SpzEncoder<T> {
         write_u32_le(&mut raw, SPZ_MAGIC);
         write_u32_le(&mut raw, version);
         write_u32_le(&mut raw, num_splats as u32);
-        raw.push(sh_degree);
+        raw.push(sh_degree as u8);
         raw.push(fractional_bits);
         let mut flags: u8 = 0;
         if flag_antialias { flags |= 0x01; }
